refactor(monitoring): split type tests for GH727
