+++ specs/GH727/tasks.md
@@ -162,24 +162,30 @@ GH-727 / #727
 - [x] `SP727-T149` Owner: coordinator. Done when: chat/message/content-part, model/function name, and image/base64/audio validation logic move into focused child modules while preserving public signature, regex patterns, error strings, and validation order. Verify: `rg -n "validate_chat_completion_request|validate_chat_message|validate_content_part|validate_model_name|validate_function_name|validate_image_url|validate_audio_format" src/utils/data/validation/request_validator.rs src/utils/data/validation/request_validator/*.rs`.
 - [x] `SP727-T150` Owner: coordinator. Done when: original inline request validator tests move to `src/utils/data/validation/request_validator/tests.rs` without assertion changes. Verify: `rg -n "test_validate_chat_completion_valid|test_validate_model_name_valid|test_validate_image_url_valid_base64|test_validate_content_part_image_valid_details|test_validate_tool_message_valid" src/utils/data/validation/request_validator/tests.rs`.
 - [x] `SP727-T151` Owner: verification owner. Done when: all touched request validator files are below U-16's 800-line ceiling and focused request validator tests pass. Verify: `wc -l src/utils/data/validation/request_validator.rs src/utils/data/validation/request_validator/*.rs`; `cargo test utils::data::validation::request_validator --lib --all-features`.
-- [ ] `SP727-T152` Owner: verification owner. Done when: formatting, SpecRail, diff check, lib all-features check, locked all-features check, default check, PR CI, and review-thread gate pass for the request validator tranche. Verify: `cargo fmt --all -- --check`; `git diff --check`; `python3 /Users/apple/Desktop/code/AI/tool/specrail/checks/check_workflow.py --repo /Users/apple/Desktop/code/AI/tool/specrail --spec-dir "$PWD/specs/GH727"`; `cargo check --lib --all-features`; `cargo check --all-features --locked`; `cargo check`; GitHub PR CI and review-thread query.
+- [x] `SP727-T152` Owner: verification owner. Done when: formatting, SpecRail, diff check, lib all-features check, locked all-features check, default check, PR CI, and review-thread gate pass for the request validator tranche. Verify: #850 PR body, green PR CI, GraphQL reviewThreads totalCount 0, merge commit `ea1a47c3286b`, and #727 update comment.
+- [x] `SP727-T153` Owner: coordinator. Done when: after the request validator tranche merges, the next #727 tranche is selected from the remaining queue with fresh tracked-file line-count evidence. Verify: at `origin/main@ea1a47c3286b`, `src/monitoring/types.rs` is 867 lines and 26 tracked Rust files remain over the U-16 ceiling.
+- [x] `SP727-T154` Owner: coordinator. Done when: the monitoring types tranche documents why this slice is a test extraction rather than a production type facade split. Verify: `git diff -- specs/GH727/product.md specs/GH727/tech.md`.
+- [x] `SP727-T155` Owner: coordinator. Done when: `src/monitoring/types.rs` keeps production monitoring type definitions and delegates tests with `#[path = "types_tests.rs"] mod tests;`. Verify: `tail -n 20 src/monitoring/types.rs`.
+- [x] `SP727-T156` Owner: coordinator. Done when: original inline monitoring type tests move to `src/monitoring/types_tests.rs` without assertion changes. Verify: `rg -n "test_request_metrics_creation|test_alert_severity_variants|test_system_metrics_serialization" src/monitoring/types_tests.rs`.
+- [x] `SP727-T157` Owner: verification owner. Done when: both touched monitoring type files are below U-16's 800-line ceiling and focused monitoring type tests pass. Verify: `wc -l src/monitoring/types.rs src/monitoring/types_tests.rs`; `cargo test monitoring::types --lib --all-features`.
+- [ ] `SP727-T158` Owner: verification owner. Done when: formatting, SpecRail, diff check, lib all-features check, locked all-features check, default check, PR CI, and review-thread gate pass for the monitoring types tranche. Verify: `cargo fmt --all -- --check`; `git diff --check`; `python3 /Users/apple/Desktop/code/AI/tool/specrail/checks/check_workflow.py --repo /Users/apple/Desktop/code/AI/tool/specrail --spec-dir "$PWD/specs/GH727"`; `cargo check --lib --all-features`; `cargo check --all-features --locked`; `cargo check`; GitHub PR CI and review-thread query.
 
 ## 并行拆分
 
-This is a serial writable lane for the request validator file family. Other #727 large-file tranches may be planned read-only in parallel, but they must not edit this branch.
+This is a serial writable lane for the monitoring types file family. Other #727 large-file tranches may be planned read-only in parallel, but they must not edit this branch.
 
 Writable ownership for this lane:
 
 - `specs/GH727/`
-- `src/utils/data/validation/request_validator.rs`
-- `src/utils/data/validation/request_validator/*.rs`
+- `src/monitoring/types.rs`
+- `src/monitoring/types_tests.rs`
 
 ## 验证
 
 - SpecRail packet review.
 - `cargo fmt --all -- --check`
 - `git diff --check`
-- `cargo test utils::data::validation::request_validator --lib --all-features`
+- `cargo test monitoring::types --lib --all-features`
 - `cargo check --lib --all-features`
 - `cargo check --all-features --locked`
 - `cargo check`
@@ -187,5 +193,5 @@ Writable ownership for this lane:
 
 ## Handoff Notes
 
-This PR is the next #727 runtime validator module split tranche and should use `Refs #727`, not `Closes #727`.
+This PR is the next #727 monitoring types test-extraction tranche and should use `Refs #727`, not `Closes #727`.
 The issue should remain open until the final scan shows no Rust files over the U-16 ceiling.
